+++ VadereSimulator/src/org/vadere/simulator/utils/scenariochecker/AbstractScenarioCheckerMessageFormatter.java
@@ -1,4 +1,4 @@
-package org.vadere.simulator.util;
+package org.vadere.simulator.utils.scenariochecker;
 
 import java.util.Arrays;
 import java.util.Comparator;