+++ VadereSimulator/src/org/vadere/simulator/utils/scenariochecker/checks/AbstractScenarioCheck.java
@@ -1,6 +1,6 @@
-package org.vadere.simulator.util.checks;
+package org.vadere.simulator.utils.scenariochecker.checks;
 
-import org.vadere.simulator.util.ScenarioCheckerMessageBuilder;
+import org.vadere.simulator.utils.scenariochecker.ScenarioCheckerMessageBuilder;
 
 public abstract class AbstractScenarioCheck implements ScenarioCheckerTest{