+++ VadereSimulator/src/org/vadere/simulator/utils/scenariochecker/checks/topography/StairTreadSanityCheck.java
@@ -1,9 +1,9 @@
-package org.vadere.simulator.util.checks.topography;
+package org.vadere.simulator.utils.scenariochecker.checks.topography;
 
-import org.vadere.simulator.util.ScenarioCheckerMessage;
-import org.vadere.simulator.util.ScenarioCheckerReason;
-import org.vadere.simulator.util.checks.AbstractScenarioCheck;
-import org.vadere.simulator.util.checks.TopographyCheckerTest;
+import org.vadere.simulator.utils.scenariochecker.ScenarioCheckerMessage;
+import org.vadere.simulator.utils.scenariochecker.ScenarioCheckerReason;
+import org.vadere.simulator.utils.scenariochecker.checks.AbstractScenarioCheck;
+import org.vadere.simulator.utils.scenariochecker.checks.TopographyCheckerTest;
 import org.vadere.state.scenario.Stairs;
 import org.vadere.state.scenario.Topography;
 
@@ -20,8 +20,8 @@ public class StairTreadSanityCheck extends AbstractScenarioCheck implements Topo
 						.topographyWarning()
 						.target(stairs)
 						.reason(ScenarioCheckerReason.STAIRS_TREAD_DIM_WRONG
-								, "(" + Stairs.MIN_TREAD_DEPTH + "m  &lt; treadDepth  &gt; " + Stairs.MAX_TREAD_DEPTH +
-										"m) current treadDepth is: " + String.format("%.3fm", stairs.getTreadDepth()))
+								, "[" + Stairs.MIN_TREAD_DEPTH + "m - " + Stairs.MAX_TREAD_DEPTH +
+										"m] current treadDepth is: " + String.format("%.3fm", stairs.getTreadDepth()))
 						.build()
 				);
 			}